Add black and white list support
This patch adds the black and white list support for hub
mirror action.

The back list has **high priority**, that mean if the repo
in the back list will never mirror, even the repo is also
in the white list.

And the white list can be used as single repo mirror, you
can set white_list as specific repo to only sync this single
repo.

Closes: #19, #14
Related: #18